Added clarification note for custom sinks
In a Kubernetes environment where I had set specific memory limits, using a "scoped" service much like the one above, it was leaking the `Logger` object with each http request to the IdentityServer.  Eventually this would result in my pod getting "OOMKilled" with an error code of 137 - memory limit exceeded.  By changing the service registration to a singleton the memory is staying steady.